Paginated stocks, fixed clients issue

+++ DevSync.PocPro/shops/shared/DevSync.PocPro.Shops.Shared/Dtos/PagedResponse.cs
@@ -2,7 +2,7 @@ namespace DevSync.PocPro.Shops.Shared.Dtos;
 
 public class PagedResponse<TResponse>  where TResponse : class
 {
-    public PagedResponse(int page, int pageSize, int totalCount, IEnumerable<TResponse> items)
+    public PagedResponse(int page, int pageSize, long totalCount, IEnumerable<TResponse> items)
     {
         Page = page;
         PageSize = pageSize;
@@ -12,7 +12,7 @@ public class PagedResponse<TResponse>  where TResponse : class
 
     public int Page { get; }
     public int PageSize { get; }
-    public int TotalCount { get; }
+    public long TotalCount { get; }
     public IEnumerable<TResponse> Items { get; }
     public bool HasPrev => Page > 1;
     public int TotalPages => (int)Math.Ceiling((double)TotalCount / PageSize);
